trimmed first 15 seconds

+++ ndmg/preproc/preproc_func.py
@@ -114,7 +114,7 @@ class preproc_func():
         # of radiofrequency excitation
         func_im = nb.load(self.func)
         tr = func_im.header.get_zooms()[3]
-        nvol_trim = np.floor(15/float(tr))
+        nvol_trim = int(np.floor(15/float(tr)))
         # remove the first nvol_trim timesteps
         trimmed_dat = func_im.get_data()[:,:,:, nvol_trim:]
         trimmed_im = nb.Nifti1Image(dataobj=trimmed_dat, header=func_im.header,
